% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_cea_metric.R
\name{cea_metric}
\alias{cea_metric}
\title{Simple wrapper function to create a raster stack with network cea metric per species}
\usage{
cea_metric(path, metric)
}
\description{
Simple wrapper function to create a raster stack with network cea metric per species
}
