% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_utils.R
\name{global_parameters}
\alias{global_parameters}
\title{Utility functions for the research compendium}
\usage{
global_parameters()
}
\arguments{
\item{did}{unique identifiers of raw data to be loaded}

\item{name}{names of ingrid or output data to be loaded}

\item{type}{type of data to load, one of "raw", "ingrid", "output"}
}
\description{
Utility functions for the research compendium
}
\section{Functions}{
\itemize{
\item \code{global_parameters()}: load research compendium global parameters and add as \code{global_param} object in global environment

}}
