% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_metanetwork.R
\name{coordCircle}
\alias{coordCircle}
\title{Let's begin by creating a function that will give us the x and y coordinates
of the outside of a circle given a certain radius}
\usage{
coordCircle(theta = NULL, radius = 1)
}
\description{
Let's begin by creating a function that will give us the x and y coordinates
of the outside of a circle given a certain radius
}
