% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_cumulative_risk.R
\name{CumulativeRisk}
\alias{CumulativeRisk}
\title{Assessment of network-scale risk for species}
\usage{
CumulativeRisk(focusID, biotic, drivers, vulnerability, sensitivity)
}
\description{
Assessment of network-scale risk for species
}
