% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_atlas.R
\name{fig_atlas}
\alias{fig_atlas}
\title{Export figures for atlas}
\usage{
fig_atlas(
  type = c("footprint", "cea", "difference", "metanetwork", "contribution", "species",
    "drivers")
)
}
\description{
Export figures for atlas
}
