% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_contribution_group.R
\name{fig_contribution_group_}
\alias{fig_contribution_group_}
\alias{fig_contribution_group}
\title{Export figures for stressor contribution to direct and indirect effects per period}
\usage{
fig_contribution_group_()

fig_contribution_group(cekm, period)
}
\description{
Export figures for stressor contribution to direct and indirect effects per period

Export figures for stressor contribution to direct and indirect effects
}
\section{Functions}{
\itemize{
\item \code{fig_contribution_group_()}: contribution stacked barplot per period

\item \code{fig_contribution_group()}: contribution stacked barplot

}}
