% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_metanetwork.R, R/fig_metanetwork_nceahab.R
\name{fig_metanetwork_}
\alias{fig_metanetwork_}
\alias{fig_metanetwork}
\alias{fig_metanetwork_nceahab_}
\alias{fig_metanetwork_nceahab}
\title{Export figures for metanetwork}
\usage{
fig_metanetwork_()

fig_metanetwork(
  cekm,
  bg = "#ffffff",
  out = here::here("figures", "metanetwork"),
  nm = "metanetwork.png"
)

fig_metanetwork_nceahab_()

fig_metanetwork_nceahab(
  bg = "#ffffff",
  out = here::here("figures", "metanetwork"),
  nm = "metanetwork.png"
)
}
\description{
Export figures for metanetwork

Export figures for metanetwork

Export figures for metanetwork

Export figures for metanetwork
}
\section{Functions}{
\itemize{
\item \code{fig_metanetwork_()}: metanetwork per period

\item \code{fig_metanetwork()}: metanetwork figure

\item \code{fig_metanetwork_nceahab_()}: metanetwork per period

\item \code{fig_metanetwork_nceahab()}: metanetwork figure

}}
