% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_modules.R
\name{format_modules}
\alias{format_modules}
\title{Format module data and export as RData for assessment. This should be changed in the future.}
\usage{
format_modules()
}
\description{
Format module data and export as RData for assessment. This should be changed in the future.
}
