% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_bib.R
\name{gather_bib}
\alias{gather_bib}
\title{Script to create a single bib file from the downloaded raw data using pipedat}
\usage{
gather_bib()
}
\description{
Script to create a single bib file from the downloaded raw data using pipedat
}
