% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bib.R
\name{get_bib}
\alias{get_bib}
\title{Script to create a single bib file from the downloaded raw data using pipedat}
\usage{
get_bib()
}
\description{
Script to create a single bib file from the downloaded raw data using pipedat
}
