% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0001.R
\name{get_data0001}
\alias{get_data0001}
\title{Data 0001 : Federal Marine Bioregions}
\source{
https://open.canada.ca/data/en/dataset/23eb8b56-dac8-4efc-be7c-b8fa11ba62e9
}
\usage{
get_data0001()
}
\description{
The spatial planning framework for Canada's national network of Marine Protected Areas (MPAs) is comprised of 13 ecologically defined bioregions that cover Canada's oceans and the Great Lakes.
}
\details{
This function loads and formats the data
}
\keyword{}
