% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0002-study_grid.R
\name{get_data0002}
\alias{get_data0002}
\title{Data 0002 : Study grid}
\source{
TO DO
}
\usage{
get_data0002()
}
\description{
Gridded study area used jointly with N. Kelly and G. Murphy for the Maritimes region
}
\details{
Cette fonction formatte les données
}
\keyword{grid}
\keyword{study}
