% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0007-conservation_measures.R
\name{get_data0007}
\alias{get_data0007}
\title{Data 0007 : Other Effective Area-Based Conservation Measures}
\source{
https://open.canada.ca/data/en/dataset/44769543-7a23-4991-a53f-c2cf7c7a946f
}
\usage{
get_data0007()
}
\description{
This dataset contains area-based management measures that constitute 'other effective area-based conservation measures' ('other measures') according to DFO's Operational Guidance for Identifying ‘Other Effective Area-Based Conservation Measures’ in Canada’s Marine Environment.
}
\details{
This function loads and formats the data
}
\keyword{areas}
\keyword{protected}
