% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0008-sea_level.R
\name{get_data0008}
\alias{get_data0008}
\title{Data 0008 : CanCoast 2.0}
\source{
https://open.canada.ca/data/en/dataset/73714ed4-a795-a7ae-7e93-36100ce7c242
}
\usage{
get_data0008()
}
\description{
CanCoast 2.0: data and indices to describe the sensitivity of Canada's marine coasts to changing climate
}
\details{
This function loads and formats the data
}
\keyword{change}
\keyword{climate}
\keyword{level}
\keyword{sea}
\keyword{stressors}
