% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0009-marine_seismic.R
\name{get_data0009}
\alias{get_data0009}
\title{Data 0009 : Canadian National Marine Seismic Data Repository}
\source{
https://ouvert.canada.ca/data/fr/dataset/e1fa0090-4b06-e476-5c71-e2326666a4d0
}
\usage{
get_data0009()
}
\description{
The Geological Survey of Canada (Atlantic and Pacific) has collected marine survey field records on marine expeditions for over 50 years. This release makes available the results of an ongoing effort to scan and convert our inventory of analog marine survey field records (seismic, sidescan and sounder) to digital format.
}
\details{
This function loads and formats the data
}
\keyword{stressors}
