% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0010-fisheries_atlas.R
\name{get_data0010}
\alias{get_data0010}
\title{Data 0010 : Maritimes Region Fisheries Atlas: Catch Weight Landings Mapping (2014–2018)}
\source{
https://open.canada.ca/data/en/dataset/44ef4d33-20b7-45fc-974c-d73a0a8fbae8

https://waves-vagues.dfo-mpo.gc.ca/Library/40885690.pdf
}
\usage{
get_data0010()
}
\description{
These datasets show commercial fisheries catch weight landings of directed fisheries and bycatch from the Scotian Shelf, the Bay of Fundy, and Georges Bank from NAFO Divisions 4VWX and the Canadian portions of 5Y and 5Z
}
\details{
This function loads and formats the data
}
\keyword{commercial}
\keyword{fisheries}
