% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0011-contaminated_sites.R
\name{get_data0011}
\alias{get_data0011}
\title{Data 0011 : Federal Contaminated Sites Inventory}
\source{
https://www.tbs-sct.gc.ca/fcsi-rscf/home-accueil-eng.aspx

https://www.tbs-sct.gc.ca/fcsi-rscf/numbers-numeros-eng.aspx?qid=369592 (New-Brunswick)

https://www.tbs-sct.gc.ca/fcsi-rscf/numbers-numeros-eng.aspx?qid=369577 (Nova-Scotia)

https://www.tbs-sct.gc.ca/fcsi-rscf/numbers-numeros-eng.aspx?qid=671230 (Newfoundland)

https://www.tbs-sct.gc.ca/fcsi-rscf/numbers-numeros-eng.aspx?qid=671235 (Prince Edward Island)

https://www.tbs-sct.gc.ca/fcsi-rscf/numbers-numeros-eng.aspx?qid=671240 (Quebec)

https://open.canada.ca/data/en/dataset/1d42f7b9-1549-40aa-8ac6-0e0302ff2902 (full dataset, but in XML format)
}
\usage{
get_data0011()
}
\description{
The Federal Contaminated Sites Inventory includes information on all known federal contaminated sites under the custodianship of departments, agencies and consolidated Crown corporations as well as those that are being or have been investigated to determine whether they have contamination arising from past use that could pose a risk to human health or the environment. The inventory also includes non-federal contaminated sites for which the Government of Canada has accepted some or all financial responsibility. It does not include sites where contamination has been caused by, and which are under the control of, enterprise Crown corporations, private individuals, firms or other levels of government.
}
\details{
This function loads and formats the data
}
\keyword{contaminated}
\keyword{sites}
\keyword{stressors}
