% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataXXXX-marine_disposal_sites.R
\name{get_data0012}
\alias{get_data0012}
\title{Data 0012 : Marine disposal sites}
\source{
https://open.canada.ca/data/en/dataset/da99526e-284f-4e06-8d04-193785cd1a96)
}
\usage{
get_data0012()
}
\description{
The Active and Inactive Disposal at Sea Sites in Canadian Waters dataset provides spatial and related information of at-sea disposal sites approved for use in Canada in the last ten years and that remain open for consideration for additional use.
}
\details{
This function loads and formats the data
}
\keyword{}
