% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0013-hardened_shoreline.R
\name{get_data0013}
\alias{get_data0013}
\title{Data 0013: Atlantic Shoreline Classification}
\source{
https://open.canada.ca/data/en/dataset/30449352-2556-42df-9ffe-47ea8e696f91
}
\usage{
get_data0013()
}
\description{
The Atlantic dataset is part of Environment and Climate Change Canada’s Shoreline Classification and Pre-Spill database. Shoreline classification data has been developed for use by the Environmental Emergencies Program of Environment and Climate Change Canada for environmental protection purposes. Marine and estuarine shorelines are classified according to the character (substrate and form) of the upper intertidal (foreshore) or upper swash zone (Sergy, 2008).
}
\details{
This function loads and formats the data
}
\keyword{}
