% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0014-carms_checklist.R
\name{get_data0014}
\alias{get_data0014}
\title{Data 0014: WoRMS North-West Atlantic species checklist}
\source{
http://www.marinespecies.org/carms/aphia.php?p=checklist&action=search&gu_id=8048&tRank=220&inc_sub=1&status=pv
}
\usage{
get_data0014()
}
\description{
CaRMS and WoRMS species list for the North-West Atlantic area of interest
}
\keyword{}
