% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataXXXX-Conservation_measures.R,
%   R/dataXXXX-Fishing_atlas.R, R/dataXXXX-Fishing_lobster.R,
%   R/dataXXXX-Fishing_longline_trap.R, R/dataXXXX-MPA.R,
%   R/dataXXXX-NS_spatial_data.R, R/dataXXXX-Oil_gas.R, R/dataXXXX-Shipping.R,
%   R/dataXXXX-marine_seismic.R
\name{get_dataXXXX}
\alias{get_dataXXXX}
\title{Data XXXX : Other Effective Area-Based Conservation Measures}
\source{
https://open.canada.ca/data/en/dataset/44769543-7a23-4991-a53f-c2cf7c7a946f

https://open.canada.ca/data/en/dataset/44ef4d33-20b7-45fc-974c-d73a0a8fbae8

https://waves-vagues.dfo-mpo.gc.ca/Library/40885690.pdf

https://open.canada.ca/data/en/dataset/64f741d7-1129-49dd-9e5c-2b1de79024f0

https://waves-vagues.dfo-mpo.gc.ca/Library/40573230.pdf

https://open.canada.ca/data/en/dataset/3d2e1a84-20f5-4a61-90e4-94760c80ebb9

http://publications.gc.ca/collections/collection_2019/mpo-dfo/Fs97-6-3293-eng.pdf

https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250

https://nsgi.novascotia.ca/gdd/

https://novascotia.ca/fish/aquaculture/site-mapping-tool/

https://www.cnsopb.ns.ca/what-we-do/lands-management/maps-and-coordinates

https://callforbids.ca/data-environment/gis-information  (See link to “General Shapefiles”)

https://callforbids.ca/geoscience-and-data

http://publications.gc.ca/collections/collection_2012/mpo-dfo/Fs97-6-2966-eng.pdf

https://ouvert.canada.ca/data/fr/dataset/e1fa0090-4b06-e476-5c71-e2326666a4d0
}
\usage{
get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()

get_dataXXXX()
}
\description{
Oceans Act Marine Protected Areas

Fisheries Atlas: Catch Weight by Species or Species Grouping on a Hexagon Grid (2014-2018):

Fisheries Atlas: Maritimes Region Inshore Lobster (2012-2014) (WARNING: check, now in the process of being updated to 2015-2019)

Fisheries Atlas: Maritimes Region Inshore Lobster (2012-2014) (WARNING: check, now in the process of being updated to 2015-2019)

Oceans Act Marine Protected Areas

Offshore oil and gas, it's probably best to get data from CNSOPB.
}
\details{
This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data

This function loads and formats the data
}
\keyword{}
