% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_individual_risk.R
\name{IndividualRisk}
\alias{IndividualRisk}
\title{Assessment of species-scale risk for species}
\usage{
IndividualRisk(drivers, vulnerability, sensitivity)
}
\description{
Assessment of species-scale risk for species
}
