% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_aoi.R
\name{int_aoi}
\alias{int_aoi}
\title{Area of interest}
\usage{
int_aoi()
}
\description{
Area of interest for contribution agreement on the assessment of network cumulative effects of global changes on communities of the Scotian Shelf Bioregion
}
\details{
This function exports the area of interest for the project
}
\keyword{area}
\keyword{interest}
\keyword{of}
\keyword{scotian}
\keyword{shelf}
