% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_habitats_sensitivity.R
\name{make_habitats_sensitivity}
\alias{make_habitats_sensitivity}
\title{Prepare habitat sensitivity data}
\usage{
make_habitats_sensitivity()
}
\description{
Prepare habitat sensitivity data
}
