% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_sensitivity.R
\name{make_species_sensitivity}
\alias{make_species_sensitivity}
\title{Species-specific sensitivity data to use for assessment}
\usage{
make_species_sensitivity()
}
\description{
Species-specific sensitivity data to use for assessment
}
