% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trophic_sensitivity.R
\name{make_trophic_sensitivity}
\alias{make_trophic_sensitivity}
\title{Trophic sensitivity data to use for assessment}
\usage{
make_trophic_sensitivity()
}
\description{
Trophic sensitivity data to use for assessment
}
