% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_network_risk.R
\name{network_metric}
\alias{network_metric}
\title{Assessment of network-scale risk for species not involved in an interaction
Simple wrapper function to create a raster stack with network cea metric per species}
\usage{
network_metric(metric)
}
\description{
Assessment of network-scale risk for species not involved in an interaction
Simple wrapper function to create a raster stack with network cea metric per species
}
