% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_network_risk.R
\name{network_risk}
\alias{network_risk}
\title{Assessment of network-scale risk for species}
\usage{
network_risk(focusID, biotic, drivers, vulnerability, sensitivity)
}
\description{
Assessment of network-scale risk for species
}
