% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_network_risk.R
\name{network_risk_disconnect}
\alias{network_risk_disconnect}
\title{Assessment of network-scale risk for species not involved in an interaction}
\usage{
network_risk_disconnect(drivers, vulnerability, sensitivity)
}
\description{
Assessment of network-scale risk for species not involved in an interaction
}
