% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cea_km2_ori.R
\name{out_cea_km2_}
\alias{out_cea_km2_}
\title{Function to export cumulative drivers and species (richness)}
\usage{
out_cea_km2_()
}
\description{
Function to export cumulative drivers and species (richness)
}
