% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_difference.R
\name{out_difference}
\alias{out_difference}
\title{Function to assess difference between assessments in both periods}
\usage{
out_difference()
}
\description{
Function to assess difference between assessments in both periods
}
