% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_exposure.R
\name{out_exposure}
\alias{out_exposure}
\title{Function to export cumulative stressors and species (richness)}
\usage{
out_exposure()
}
\description{
Function to export cumulative stressors and species (richness)
}
