% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_exposure.R, R/out_footprint.R
\name{out_footporint}
\alias{out_footporint}
\title{Function to export cumulative stressors and species (richness)}
\usage{
out_footporint()

out_footporint()
}
\description{
Function to export cumulative stressors and species (richness)

Function to export cumulative stressors and species (richness)
}
