% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_footprint.R
\name{out_footprint}
\alias{out_footprint}
\title{Function to export cumulative drivers and species (richness)}
\usage{
out_footprint()
}
\description{
Function to export cumulative drivers and species (richness)
}
