% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_indirect_species_habitats.R
\name{out_indirect_species_habitats_}
\alias{out_indirect_species_habitats_}
\title{Assessing indirect effects of environmental stressors to individual species arising from the direct effects of stressors to habitats and species habitat use:}
\usage{
out_indirect_species_habitats_()
}
\arguments{
\item{biotic}{rasters of binary species distribution}

\item{habitats}{rasters of binary habitat distribution}

\item{habitats_cea}{rasters of cumulative effects to habitats (Halpern method)}

\item{potential_habitats}{matrix \link{species, habitats} with potential habitat use identified as 0 and 1. This is done to consider the vertical use within the water column.}
}
\description{
Assessing indirect effects of environmental stressors to individual species arising from the direct effects of stressors to habitats and species habitat use:
}
