% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_nceadfo.R
\name{plot_legend_cont}
\alias{plot_legend_cont}
\alias{plot_legend_bin}
\alias{plot_legend_dual}
\title{Plot legend}
\usage{
plot_legend_cont(
  range = c(0, 1),
  pal = NULL,
  cexMain = 0.75,
  cexSub = 0.5,
  minUp = 0.055,
  mainTitle = NULL,
  subTitle = NULL,
  n = 5,
  colText = "#dedede"
)

plot_legend_bin(
  col,
  cexMain = 0.75,
  cexSub = 0.5,
  minUp = 0.055,
  mainTitle = NULL,
  subTitle = NULL,
  colText = "#dedede"
)

plot_legend_dual(
  rMax,
  pal1 = NULL,
  pal2 = NULL,
  cexMain = 0.75,
  cexSub = 0.5,
  minUp = 0.055,
  mainTitle = NULL,
  subTitle = NULL,
  n = 5,
  colText = "#dedede"
)
}
\arguments{
\item{range}{numeric, vector with minimal and maximal values}

\item{pal}{character, vector of colors, or color palette}

\item{minUp}{numeric, minimum upper side to write as a function of bbox extent}

\item{mainTitle}{character, type of legend. Choices are 'continuous', 'binary' or 'categorical'}

\item{subTitle}{character, type of legend. Choices are 'continuous', 'binary' or 'categorical'}

\item{colText}{color of text}

\item{pal1}{character, vector of colors, or color palette for first palette used in dual}

\item{pal2}{character, vector of colors, or color palette for second palette used in dual}

\item{cex.text}{numeric, cex for legend text}

\item{type}{character, type of legend. Choices are 'continuous', 'binary' or 'categorical'}

\item{nTick}{numeric, number of ticks in the legend}
}
\value{
Opens a graphical interface with the plot
}
\description{
Function to create a legend
}
\keyword{legend}
\keyword{plot,}
