% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_nceadfo.R
\name{plot_nceadfo}
\alias{plot_nceadfo}
\alias{plot_nceadfo.stars}
\alias{plot_nceadfo_dual}
\title{plot nceadfo data maps}
\usage{
plot_nceadfo(dat, ...)

\method{plot_nceadfo}{stars}(dat)

plot_nceadfo_dual(dat)
}
\arguments{
\item{dat}{object of class stars}

\item{...}{further specifications, see \link{plot} and details.}

\item{main}{main title}

\item{type}{secondary title / data type (see metadata files)}

\item{subtitle}{subtitle}

\item{unit_data}{units of data}

\item{references}{data citation used for integrated data}

\item{city}{logical, if TRUE name of cities are added to graph}

\item{minUp}{numeric, minimum upper side to write as a function of bbox extent (legend)}
}
\description{
base plot functions for nceadfo project
}
\section{Functions}{
\itemize{
\item \code{plot_nceadfo_dual}: plot with dual legend color centered on 0
}}

\examples{
 load_integrated("navigation")
 plot_nceadfo(navigation[,10], "Navigation", "Recherche gouvernementale")

}
