% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_plot_nceadfo.R
\name{plot_nceadfo_simple}
\alias{plot_nceadfo_simple}
\title{Function for simple plots for presentations}
\usage{
plot_nceadfo_simple(
  dat,
  legend = TRUE,
  land = TRUE,
  mainTitle = NULL,
  subTitle = NULL,
  range = NULL
)
}
\description{
Function for simple plots for presentations
}
