% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_metanetwork.R
\name{plotMetanetwork}
\alias{plotMetanetwork}
\title{Function to plot metanetwork using hierarchical edge bundling}
\usage{
plotMetanetwork(
  metanetwork,
  rad1 = 0.925,
  rad2 = 1,
  sizeEdge = T,
  colPal = pal_insileco,
  type = "all",
  focus = NULL,
  colLinks = "#876b40",
  colShadow = "#f4f4f4",
  shadowEdge = T,
  cexSize = NULL
)
}
\description{
Function to plot metanetwork using hierarchical edge bundling
}
