% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_species_risk.R
\name{species_risk}
\alias{species_risk}
\title{Assessment of species-scale risk for species}
\usage{
species_risk(drivers, vulnerability, sensitivity)
}
\description{
Assessment of species-scale risk for species
}
