% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_commercial_fisheries.R
\name{st_commercial_fisheries}
\alias{st_commercial_fisheries}
\title{Commercial fisheries}
\usage{
st_commercial_fisheries()
}
\description{
Integrated data for commercial fisheries in the Scotian Shelf bioregion
}
\details{
This function imports formatted data and prepares the data layer that will be used for the cumualtive effects assessment
}
\keyword{commercial}
\keyword{fisheries}
\keyword{stressors}
