% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_contaminated_sites.R
\name{st_contaminated_sites}
\alias{st_contaminated_sites}
\title{Contaminated sites}
\usage{
st_contaminated_sites()
}
\description{
Integrated data for contaminated sites in the Scotian Shelf bioregion
}
\details{
This function imports formatted data and prepares the data layer that will be used for the cumualtive effects assessment
}
\keyword{contaminated}
\keyword{sites}
\keyword{stressors}
