% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_hardened_shoreline.R
\name{st_hardened_shoreline}
\alias{st_hardened_shoreline}
\title{Hardened shoreline}
\usage{
st_hardened_shoreline()
}
\description{
Integrated data for hardened shoreline in the Scotian Shelf bioregion
}
\details{
This function imports formatted data and prepares the data layer that will be used for the cumualtive effects assessment
}
\keyword{hardened}
\keyword{shoreline}
\keyword{stressors}
