% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_inorganic_input.R
\name{st_inorganic_input}
\alias{st_inorganic_input}
\title{Inorganic input}
\usage{
st_inorganic_input()
}
\description{
Integrated data for inorganic input in the Scotian Shelf bioregion
}
\details{
This function imports formatted data and prepares the data layer that will be used for the cumualtive effects assessment
}
\keyword{contaminated}
\keyword{inorganic}
\keyword{pollution}
\keyword{sites}
\keyword{stressors}
