% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_st_marine_debris.R
\name{st_marine_debris}
\alias{st_marine_debris}
\title{Marine debris}
\usage{
st_marine_debris()
}
\description{
Integrated data for marine debris in the Scotian Shelf bioregion
}
\details{
This function imports formatted data and prepares the data layer that will be used for the cumualtive effects assessment
}
\keyword{debris}
\keyword{disposal}
\keyword{dumping}
\keyword{marine}
\keyword{ocean}
\keyword{sites}
\keyword{stressors}
