% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_contribution_nceahab.R,
%   R/fig_contribution_taxa.R
\name{fig_contribution_taxa_}
\alias{fig_contribution_taxa_}
\alias{fig_contribution_nceahab}
\alias{fig_contribution_taxa}
\title{Export figures for stressor contribution to direct and indirect effects per period}
\usage{
fig_contribution_taxa_()

fig_contribution_nceahab(direct, indirect, period)

fig_contribution_taxa_()

fig_contribution_taxa(
  direct,
  indirect,
  period,
  out = here::here("figures", "contribution")
)
}
\description{
Export figures for stressor contribution to direct and indirect effects per period

Export figures for stressor contribution to direct and indirect effects

Export figures for stressor contribution to direct and indirect effects per period

Export figures for stressor contribution to direct and indirect effects
}
\section{Functions}{
\itemize{
\item \code{fig_contribution_taxa_()}: contribution stacked barplot per period

\item \code{fig_contribution_nceahab()}: contribution stacked barplot

\item \code{fig_contribution_taxa_()}: contribution stacked barplot per period

\item \code{fig_contribution_taxa()}: contribution stacked barplot

}}
