% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_metanetwork.R
\name{fig_metanetwork_}
\alias{fig_metanetwork_}
\alias{fig_metanetwork}
\title{Export figures for metanetwork}
\usage{
fig_metanetwork_()

fig_metanetwork(cekm, period, type)
}
\description{
Export figures for metanetwork

Export figures for metanetwork
}
\section{Functions}{
\itemize{
\item \code{fig_metanetwork_}: metanetwork per period

\item \code{fig_metanetwork}: metanetwork figure
}}

