% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data0005-commercial_fisheries.R
\name{get_data0005}
\alias{get_data0005}
\title{Data 0005 : Species index for commercial fisheries (id: 0003)}
\source{
Fisheries and Oceans Canada (2021). Index of species in departement of Fisheries and Oceans Canada’s Fisheries and Oceans Canada Zonal Interchange File Format (ZIFF) data. A compilation of landing data from logbook data between 2000 and 2020. \link{Data accessed 2021-07-15}
}
\usage{
get_data0005()
}
\description{
Species index for Zonal Interchange File Format (ZIFF) data between 2000 and 2020
}
\details{
Cette fonction formatte les données
}
\keyword{commercial}
\keyword{fisheries}
\keyword{stressors}
