% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cea_km2.R
\name{out_cea_km2}
\alias{out_cea_km2}
\title{Function to export cumulative drivers and species (richness)}
\usage{
out_cea_km2()
}
\description{
Function to export cumulative drivers and species (richness)
}
