% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cea_km2.R
\name{out_cea_km2}
\alias{out_cea_km2}
\title{Script to extract cumulative effects scores / km2 for each species from the species and network-scale assessments}
\usage{
out_cea_km2()
}
\description{
Script to extract cumulative effects scores / km2 for each species from the species and network-scale assessments
}
